Change ids to long or uuid

+++ src/main/kotlin/at/markusnentwich/helicon/dto/StateDto.kt
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 @Schema(name = "State")
 data class StateDto(
         @Schema(description = "the id of this state", example = "73")
-        val id: Int?,
+        val id: Long?,
         @Schema(description = "the name of this state", example = "Österreich")
         val name: String,
         @Schema(description = "the zone where this states belongs to")
@@ -15,7 +15,7 @@ data class StateDto(
 @Schema(name = "Zone")
 data class ZoneDto(
         @Schema(description = "the id of this zone", example = "73")
-        val id: Int?,
+        val id: Long?,
         @Schema(description = "the name of this zone", example = "EU")
         val name: String,
         @Schema(description = "the shipping of this zone in subunits (e.g. cent)", example = "700")
