+++ src/main/kotlin/at/markusnentwich/helicon/services/MetaService.kt
@@ -31,7 +31,7 @@ interface MetaService {
                 ApiResponse(responseCode = "404", description = "no state found with such id")
             ]
     )
-    fun getStateById(@Parameter(description = "the id of the state") @PathVariable id: Int): ResponseEntity<StateDto>
+    fun getStateById(@Parameter(description = "the id of the state") @PathVariable id: Long): ResponseEntity<StateDto>
 
     @RequestMapping("/state", method = [RequestMethod.POST])
     @Operation(summary = "create a new state")
@@ -59,7 +59,7 @@ interface MetaService {
             ]
     )
     fun updateState(@Parameter(description = "the state to update") @RequestBody state: StateDto,
-                    @Parameter(description = "id of the state") @PathVariable id: Int): ResponseEntity<StateDto>
+                    @Parameter(description = "id of the state") @PathVariable id: Long): ResponseEntity<StateDto>
 
     @RequestMapping("/state/{id}", method = [RequestMethod.DELETE])
     @Operation(summary = "deletes an existing state")
@@ -72,7 +72,7 @@ interface MetaService {
                 ApiResponse(responseCode = "404", description = "no state with such id found")
             ]
     )
-    fun deleteState(@Parameter(description = "the id of the state") @PathVariable id: Int)
+    fun deleteState(@Parameter(description = "the id of the state") @PathVariable id: Long)
 
     @RequestMapping("/zone", method = [RequestMethod.GET])
     @Operation(summary = "returns all zones")
@@ -89,7 +89,7 @@ interface MetaService {
                 ApiResponse(responseCode = "404", description = "no zone with such id found")
             ]
     )
-    fun getZoneById(@Parameter(description = "the id of the zone") @PathVariable id: Int): ResponseEntity<ZoneDto>
+    fun getZoneById(@Parameter(description = "the id of the zone") @PathVariable id: Long): ResponseEntity<ZoneDto>
 
     @RequestMapping("/zone", method = [RequestMethod.POST])
     @Operation(summary = "insert a new zone")
@@ -116,7 +116,7 @@ interface MetaService {
             ]
     )
     fun updateZone(@Parameter(description = "the zone to update") @RequestBody zone: ZoneDto,
-                   @Parameter(description = "id of the zone") @PathVariable id: Int): ResponseEntity<ZoneDto>
+                   @Parameter(description = "id of the zone") @PathVariable id: Long): ResponseEntity<ZoneDto>
 
     @RequestMapping("/zone/{id}", method = [RequestMethod.DELETE])
     @Operation(summary = "delete an existing zone")
@@ -129,5 +129,5 @@ interface MetaService {
                 ApiResponse(responseCode = "404", description = "no zone with such id found")
             ]
     )
-    fun deleteZone(@Parameter(description = "the id of the zone") @PathVariable id: Int)
+    fun deleteZone(@Parameter(description = "the id of the zone") @PathVariable id: Long)
 }
